Fix complier warning

+++ Flickr/NetworkService/FlickrPhotoNetworkService.m
@@ -15,7 +15,7 @@
                     completionBlock:(void (^)(NSDictionary *filmData))completionBlock {
   NSURLSessionConfiguration *configuration = [NSURLSessionConfiguration defaultSessionConfiguration];
   NSURLSession *session = [NSURLSession sessionWithConfiguration:configuration];
-  NSURL *url = [NSURL URLWithString:[NSString stringWithFormat:@"https://api.flickr.com/services/rest/?method=flickr.photos.search&api_key=3e7cc266ae2b0e0d78e279ce8e361736&format=json&nojsoncallback=1&safe_search=1&page=%lu&text=%@", pageNumber, searchString]];
+  NSURL *url = [NSURL URLWithString:[NSString stringWithFormat:@"https://api.flickr.com/services/rest/?method=flickr.photos.search&api_key=3e7cc266ae2b0e0d78e279ce8e361736&format=json&nojsoncallback=1&safe_search=1&page=%lu&text=%@", (unsigned long)pageNumber, searchString]];
   [[session dataTaskWithURL:url
           completionHandler:^(NSData *data, NSURLResponse *response, NSError *error) {
             if (error == nil && data != nil) {
